[WIP] [skip ci] Added shared logic for connections hook and connections
