+++ shared_models/shared_models.py
@@ -2,6 +2,7 @@ from enum import Enum
 from typing import Optional, Dict, Literal, Union
 
 from aries_cloudcontroller import (
+    ConnRecord,
     IndyProof,
     IndyProofRequest,
     V10PresentationExchange,
@@ -169,3 +170,27 @@ def presentation_record_to_model(
         )
     else:
         raise ValueError("Record format unknown.")
+
+
+def conn_record_to_connection(connection_record: ConnRecord):
+    return Connection(
+        connection_id=connection_record.connection_id,
+        connection_protocol=connection_record.connection_protocol,
+        created_at=connection_record.created_at,
+        invitation_mode=connection_record.invitation_mode,
+        their_role=connection_record.their_role,
+        my_did=connection_record.my_did,
+        state=connection_record.rfc23_state,
+        alias=connection_record.alias,
+        their_did=connection_record.their_did,
+        their_label=connection_record.their_label,
+        their_public_did=connection_record.their_public_did,
+        updated_at=connection_record.updated_at,
+        error_msg=connection_record.error_msg,
+        invitation_key=connection_record.invitation_key,
+        invitation_msg_id=connection_record.invitation_msg_id,
+    )
+
+
+class ConnectionsHook(HookBase, Connection):
+    pass
